requirements.txt
APScheduler==3.7.0
aio-pika==6.8.0
aioamqp==0.10.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
appdirs==1.4.4
asyncio==3.4.3
chardet==4.0.0
hawkcatcher==2.0.0
multidict==5.1.0
pika==0.10.0
pipdeptree==0.10.1
python-dotenv==0.19.2
pymongo==3.11.4
packaging==21.3
pyparsing==3.0.6
requests==2.25.1
six==1.16.0
wheel==0.29.0
yarl==1.6.3